Standardise the remaining code after the merge
